ChatCommandConverter 생성
